[refactor]: refactored ResultTableController, removed duplicates
