+++ src/main/java/com/netrunners/financialcalculator/logic/entity/ResultTableSender.java
@@ -1,5 +1,11 @@
 package com.netrunners.financialcalculator.logic.entity;
 
+import java.time.LocalDate;
+
 public interface ResultTableSender {
     void sendToResultTable();
+    String getCurrency();
+    LocalDate getStartDate();
+    LocalDate getEndDate();
+    float getBody();
 }
